A2
InputVis is my Visualization part, I have used paper.js for this.
I didn’t end up using Jquery but i have it somewhere in the code so i
added it here so it does not show an error if you try to run it.

M_Game is the code i used in particle build IDE